# Improve array access performance