@sebastienros 2023-08-28
They threw you some performance challenges ;)